requirements.txt
numpy>=1.24
pandas>=2.0
pyarrow>=14.0
scipy>=1.10
scikit-learn>=1.3
statsmodels>=0.14
matplotlib>=3.7
pyyaml>=6.0
tqdm>=4.66
torch>=2.2
pytest>=8.0
openpyxl>=3.1
